Make cli entirely extension subcommands
