+++ src/nebari/__main__.py
@@ -1,4 +1,10 @@
-from _nebari.__main__ import main
+from _nebari.cli import create_cli
+
+
+def main():
+    cli = create_cli()
+    cli()
+
 
 if __name__ == "__main__":
     main()
